pyproject.toml
[project]
name = "isla-lp-benchmark"
version = "1.1.0"
description = "LP Benchmarking Platform - Compare multiple Linear Programming solvers"
authors = [
    {name = "Oyhs-co",email = "oyhsotelo@gmail.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "gurobipy (>=13.0.1,<14.0.0)",
    "polars (>=1.39.0,<2.0.0)",
    "matplotlib (>=3.9.0,<4.0.0)",
    "numpy (>=2.4.3,<3.0.0)",
    "fpdf2 (>=2.7.0,<3.0.0)",
    "reportlab (>=4.4.10,<5.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "highspy (>=1.14.0,<2.0.0)",
    "swiglpk (>=5.0.13,<6.0.0)",
    "pulp (>=3.3.0,<4.0.0)",
    "pyscipopt (>=6.1.0,<7.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

